refactor: change stepper location
