+++ src/pages/stepper/InitialGoalMessage.jsx
